Refactor: Move cancel edit on escape to task container
